+++ app.js
@@ -15,20 +15,37 @@ app.use(express.static(path.join(__dirname, 'public')));
 const indexRouter = require('./routes/index');
 app.use('/api', indexRouter);
 
-const stationsRouter = require('./routes/stations');
-app.use('/api/stations', stationsRouter);
+require('./routes/stations');
+require('./routes/train_lines');
+require('./routes/gates');
+require('./routes/elevators');
+require('./routes/guidelines');
 
-const guidelinesRouter = require('./routes/guidelines');
-app.use('/api/guidelines', guidelinesRouter);
+// const stationsRouter = require('./routes/stations');
+// app.use('/api/stations', stationsRouter);
 
-const elevatorsRouter = require('./routes/elevators');
-app.use('/api/elevators', elevatorsRouter);
+// const trainLinesRouter = require('./routes/train_lines');
+// app.use('/api/:station_id/train_lines', trainLinesRouter);
 
-const stationsV2Router = require('./routes/stationsV2');
-app.use('/api/v2/stations', stationsV2Router);
+// const gatesRouter = require('./routes/gates');
+// app.use('/api/:station_id/train_lines/:train_line_id/gates', gatesRouter);
 
-const guidelinesV2Router = require('./routes/guidelines');
-app.use('/api/v2/guidelines', guidelinesV2Router);
+// const guidelinesRouter = require('./routes/guidelines');
+// app.use('/api:guidelines', guidelinesRouter);
+
+// const elevatorsRouter = require('./routes/elevators');
+// app.use('/api/stations/:station_id/elevators', elevatorsRouter);
+
+
+
+// const stationsRouter = require('./routes/stations');
+// app.use('/api/stations', stationsRouter);
+
+// const guidelinesRouter = require('./routes/guidelines');
+// app.use('/api/guidelines', guidelinesRouter);
+
+// const elevatorsRouter = require('./routes/elevators');
+// app.use('/api/elevators', elevatorsRouter);
 
 app.use((req, res, next) => {
     next(createError(404));
