+++ routes/index.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-    res.render('index', { title: 'Express' });
+router.get('/index', (req, res, next) => {
+    res.json({title : 'Express'});
 });
 
 module.exports = router;
